[scroll_overlay] Add a smoke-test in non-integration form, too
